# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chatbot",
        "description": "API para la gestión de usuarios, chats y mensajes de un chatbot. Documentación en español.",
        "version": "1"
    },
    "paths": {
...
        "/api/messages": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Listar mensajes paginados",
                "description": "Obtiene una lista paginada de mensajes.",
                "operationId": "4216f3eedb7be911361f1b91030253a8",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Cantidad de mensajes por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de mensajes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MessageDTO"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 15
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Enviar mensaje a un chat y obtener respuesta de IA",
                "description": "Envía un mensaje a un chat y recibe la respuesta del bot.",
                "operationId": "c85a9171473a13f151a8672235e01559",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "chat_id",
                                    "message"
                                ],
                                "properties": {
                                    "chat_id": {
                                        "description": "ID del chat",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "message": {
                                        "description": "Mensaje a enviar",
                                        "type": "string",
                                        "example": "Hola, ¿cómo estás?"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mensaje enviado y respuesta recibida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_message": {
                                            "$ref": "#/components/schemas/MessageDTO"
                                        },
                                        "bot_response": {
                                            "type": "string",
                                            "example": "¡Hola! ¿En qué puedo ayudarte hoy?"
                                        },
                                        "chat_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2024-06-25T12:34:56Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/messages/chat/{chatId}": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener mensajes por chat",
                "description": "Devuelve todos los mensajes de un chat específico.",
                "operationId": "2fb06712e46d63f0ed5b3d519aa53990",
                "parameters": [
                    {
                        "name": "chatId",
                        "in": "path",
                        "description": "ID del chat",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mensajes del chat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MessageDTO"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Chat no encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/messages/user": {
            "get": {
                "tags": [
                    "Mensajes"
                ],
                "summary": "Obtener mensajes del usuario autenticado",
                "description": "Devuelve todos los mensajes enviados por el usuario autenticado.",
                "operationId": "2f89f8f83aa49597e2140616369e5a4e",
                "responses": {
                    "200": {
                        "description": "Mensajes del usuario",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "messages": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MessageDTO"
                                            }
                                        },
                                        "count": {
                                            "type": "integer",
                                            "example": 25
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MessageDTO": {
                "title": "Mensaje",
                "description": "DTO que representa un mensaje dentro de un chat.",
                "required": [
                    "chat_id",
                    "sender_type",
                    "content",
                    "created_at"
                ],
                "properties": {
                    "chat_id": {
                        "description": "ID del chat al que pertenece el mensaje",
                        "type": "integer",
                        "example": 10
                    },
                    "sender_type": {
                        "description": "Tipo de remitente: user o bot",
                        "type": "string",
                        "example": "user"
                    },
                    "content": {
                        "description": "Contenido del mensaje",
                        "type": "string",
                        "example": "Hola, ¿cómo estás?"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del mensaje",
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-25T12:34:56Z"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Autenticación mediante token Bearer. Ingrese 'Bearer {token}' en el encabezado Authorization.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}